+++ src/frontend/apps/impress/src/components/quick-search/QuickSearchItem.tsx
@@ -3,10 +3,16 @@ import { PropsWithChildren } from 'react';
 
 type Props = {
   onSelect?: (value: string) => void;
+  id?: string;
 };
 export const QuickSearchItem = ({
   children,
   onSelect,
+  id,
 }: PropsWithChildren<Props>) => {
-  return <Command.Item onSelect={onSelect}>{children}</Command.Item>;
+  return (
+    <Command.Item value={id} onSelect={onSelect}>
+      {children}
+    </Command.Item>
+  );
 };